pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "watchdog>=6.0.0",
]
